Create README.md
In this projet we are using Deep Learning on Chest X-ray Images to Distinguish COVID-19 from Others Pneumonia Cases in addition to SARS-CoV-2 pneumonia, there are two other common types:  viral and bacterial pneumonia. They all present similar shapes and thus were difficult to distinguish. In this research, we proposed a deep, multi-layered classification of pneumonia based on a neural network using chest X-ray images. The experiment was conducted using the combination of two images from an open-source dataset collected by Cohen GP and Kagel. The data consists of 4 categories: normal pneumonia, bacterial pneumonia, viral pneumonia, and COVID19, for a total of 1,350 images. We considered two modified architectures: Xception and ResNet50. In addition, a comparison was made between the two models. The pre-trained Xception model in 100 epochs provided a classification accuracy of 94.2%.